requirements.txt
streamlit>=1.32.0
duckdb>=0.10.0
pandas>=2.0.0
pydantic>=2.0.0
openai>=1.12.0
langchain>=0.1.0
faiss-cpu>=1.7.4
pyyaml>=6.0
tabulate>=0.9.0
python-dotenv>=1.0.0
